Refactor: 머지 전 리팩토링
