--- client/src/GUI/Auth/AuthSuperFrame.java
@@ -1,12 +1,0 @@
-package GUI.Auth;
-
-import javax.swing.*;
-
-public class AuthSuperFrame  extends JFrame {
-    public AuthSuperFrame() {
-        //TODO 메인 프레임
-        setSize(1200, 800);
-
-        setVisible(true);
-    }
-}
