+++ client/src/GUI/Common/CompNames.java
@@ -1,0 +1,21 @@
+package GUI.Common;
+
+final public class CompNames {
+    //TODO: upper panel
+    public static final String searchField_u = "searchField_u";
+    public static final String searchButton_u = "searchButton_u";
+    public static final String logOutButton_u = "logOutButton_u";
+    //TODO: left panel
+    public static final String videoSearchPanel_l = "videoSearchPanel_l";
+    public static final String addToDownloadButton_l = "addToDownloadButton_l";
+
+    //TODO: right upper panel
+//    public static final String progressbar_r = "progressbar_r";
+    public static final String downloadWaitingPanel_r = "DownloadWaitingPanel_r";
+    public static final String downloadButton_r = "downloadButton_r";
+
+    //TODO: right below panel
+    public static final String downloadedListPanel_r = "downloadedListPanel_r";
+    public static final String playButton_r = "playButton_r";
+    public static final String downloadedListCell_r = "downloadedListCell_r";
+}
