+++ client/src/GUI/Common/CustomColors.java
@@ -3,7 +3,7 @@ package GUI.Common;
 import java.awt.*;
 
 public final class CustomColors {
-    public static final Color DEFAULT_GRAY = new Color(238,238,238);
+    public static final Color DEFAULT_GRAY = new Color(238, 238, 238);
 
     public static final Color MAIN_BLUE = new Color(193, 204, 222);
     public static final Color BUTTON_GREEN = new Color(136, 179, 197);
@@ -12,5 +12,3 @@ public final class CustomColors {
     public static final Color PANEL_BLUE = new Color(220, 230, 236);
     public static final Color PROGRESS_GREEN = new Color(0xff88de5b, true);
 }
-
-
