--- client/src/GUI/Download/DownloadCompNames.java
@@ -1,24 +1,0 @@
-package GUI.Download;
-
-public class DownloadCompNames {
-    //upper panel
-    public static final String searchField_u = "searchField_u";
-    public static final String searchButton_u = "searchButton_u";
-    public static final String extraButton_u = "extraButton_u";
-    public static final String logOutButton_u = "logOutButton_u";
-
-    public static final String videoSearchPanel_l = "videoSearchPanel_l";
-    public static final String addToDownloadButton_l = "addToDownloadButton_l";
-
-
-    public static final String progressbar_r = "progressbar_r";
-    public static final String downloadWaitingPanel_r = "DownloadWaitingPanel_r";
-    public static final String downloadButton_r = "downloadButton_r";
-    //left panel
-
-    //right upper panel
-
-    //right below panel
-    public static final String downloadedListPanel_r = "downloadedListPanel_r";
-    public static final String playButton_r = "playButton_r";
-}
