--- client/src/GUI/Download/DownloadTestMain.java
@@ -1,12 +1,0 @@
-package GUI.Download;
-
-import GUI.Login.LoginSuperFrame;
-
-public class DownloadTestMain {
-    public static void main(String[] args) {
-
-//        DownloadSuperFrame a = new DownloadSuperFrame();
-        LoginSuperFrame b = new LoginSuperFrame();
-
-    }
-}
