+++ client/src/GUI/Download/DownloadedListPanel.java
@@ -1,13 +1,12 @@
 package GUI.Download;
 
+import GUI.Common.CompNames;
 import GUI.Common.CustomColors;
 import dto.VideoDTO;
-
 
 import javax.swing.*;
 import java.awt.*;
 import java.util.ArrayList;
-import java.util.LinkedList;
 
 public class DownloadedListPanel extends JPanel {
     private final JLabel waitingLabel;
@@ -15,7 +14,8 @@ public class DownloadedListPanel extends JPanel {
     private final JScrollPane scrollPane;
 
     public DownloadedListPanel(ArrayList<VideoDTO> dtos) {
-        setName("downloadedListPanel_r");
+        setName(CompNames.downloadedListPanel_r);
+//        setName("downloadedListPanel_r");
         setLayout(new BorderLayout());
         setPreferredSize(new Dimension(380, 265));
         waitingLabel = new JLabel("다운로드 보관함");
@@ -31,7 +31,6 @@ public class DownloadedListPanel extends JPanel {
         for (VideoDTO dto : dtos) {
             DownloadedListCell cell = new DownloadedListCell(dto);
             contentPane.add(cell);
-            //셀간 간격
             contentPane.add(Box.createVerticalStrut(15));
         }
 
@@ -46,30 +45,30 @@ public class DownloadedListPanel extends JPanel {
     }
 
     public void updatePanel(ArrayList<VideoDTO> dtos) {
-        contentPane.removeAll(); // 이전에 추가된 모든 컴포넌트 제거
-        System.out.println(" dtos의 길이는 다음과 같습니다: "  +dtos.size());
+        contentPane.removeAll();
+
+//        System.out.println(" dtos의 길이는 다음과 같습니다: "  +dtos.size());
         for (VideoDTO dto : dtos) {
             if (dtos.size() == 1) {
-                System.out.println("셀이 하나인 경우 시작합니다.");
+//                System.out.println("셀이 하나인 경우 시작합니다.");
                 DownloadedListCell cell = new DownloadedListCell(dto);
                 contentPane.add(cell);
                 JPanel dummy = new JPanel();
-                System.out.println("더미 패널이 생겼습니다.");
-                dummy.setPreferredSize(new Dimension(200,140));
+//                System.out.println("더미 패널이 생겼습니다.");
+                dummy.setPreferredSize(new Dimension(200, 140));
                 dummy.setOpaque(true);
 //                dummy.setBackground(Color.black);
                 contentPane.add(dummy);
-                System.out.println("더미 패널이 추가되었습니다.");
-            }else{
-                System.out.println("셀 업데이트를 시작합니다.");
+//                System.out.println("더미 패널이 추가되었습니다.");
+            } else {
+//                System.out.println("셀 업데이트를 시작합니다.");
                 DownloadedListCell cell = new DownloadedListCell(dto);
                 contentPane.add(cell);
                 // 셀간 간격
                 contentPane.add(Box.createVerticalStrut(15));
             }
         }
-
-        contentPane.revalidate(); // 패널을 다시 그리기 위해 호출
+        contentPane.revalidate();
         contentPane.repaint();
 
         //스크롤바 최상단 재설정
@@ -78,5 +77,4 @@ public class DownloadedListPanel extends JPanel {
 //            verticalScrollBar.setValue(0);
 //        });
     }
-
 }
