+++ client/src/GUI/Download/VideoInfoCell.java
@@ -12,20 +12,16 @@ import java.net.URL;
 
 public class VideoInfoCell extends JPanel {
     private final JLabel imageLabel;
-//    private final JLabel titleLabel;
     private final JTextArea titleArea;
     private final JLabel viewCount;
     private final JLabel uploader;
     private final JButton addToDownloadButton;
-
     private final VideoDTO dto;
 
     public VideoInfoCell(VideoDTO dto) {
         setLayout(new BoxLayout(this, BoxLayout.X_AXIS));
-//        setBackground(Color.yellow);
         this.dto = dto;
 
-        //TODO 이미지
         imageLabel = new JLabel();
         try {
             URL url = new URL(dto.getThumbnailUrl());
@@ -39,10 +35,9 @@ public class VideoInfoCell extends JPanel {
         add(Box.createHorizontalStrut(20));
         add(imageLabel);
         add(Box.createHorizontalStrut(60));
-        // 정보
+
         JPanel infoPanel = new JPanel();
         infoPanel.setLayout(new BoxLayout(infoPanel, BoxLayout.Y_AXIS));
-//        titleLabel = new JLabel("제목: " + dto.getTitle());
 
         titleArea = new JTextArea(dto.getTitle());
         titleArea.setBackground(CustomColors.DEFAULT_GRAY);
@@ -50,7 +45,7 @@ public class VideoInfoCell extends JPanel {
         titleArea.setWrapStyleWord(true);
         titleArea.setEditable(false);
         titleArea.setAlignmentX(Component.LEFT_ALIGNMENT);
-        titleArea.setFont(new Font("맑은 고딕",Font.BOLD,12));
+        titleArea.setFont(new Font(titleArea.getFont().getName(), Font.BOLD, 12));
 
         viewCount = new JLabel("조회수: " + dto.getViewCount());
         viewCount.setAlignmentX(Component.LEFT_ALIGNMENT);
@@ -63,9 +58,8 @@ public class VideoInfoCell extends JPanel {
         infoPanel.add(titleArea);
         infoPanel.add(viewCount);
         infoPanel.add(uploader);
-        infoPanel.add(Box.createVerticalStrut(30)); // 수직 간격
+        infoPanel.add(Box.createVerticalStrut(30));
 
-        // 재생 버튼
         addToDownloadButton = new JButton("다운 목록에 추가");
         addToDownloadButton.setName("addToDownloadButton_l");
         addToDownloadButton.setPreferredSize(new Dimension(100, 30));
@@ -74,11 +68,10 @@ public class VideoInfoCell extends JPanel {
 
         addToDownloadButton.addActionListener(e -> {
             ViewController.downloadWaitingList.add(dto);
-            System.out.println("다운로드 대기리스트에 dto가 추가되었습니다.");
+//            System.out.println("다운로드 대기리스트에 dto가 추가되었습니다.");
             ViewController.updateDownloadWaitingPanel();
         });
 
         add(infoPanel);
     }
 }
-
