+++ client/src/GUI/Download/VideoSearchPanel.java
@@ -1,7 +1,6 @@
 package GUI.Download;
 
 
-
 import dto.VideoDTO;
 
 import javax.swing.*;
@@ -16,38 +15,35 @@ public class VideoSearchPanel extends JPanel {
         setName("videoSearchPanel_l");
         setLayout(new BorderLayout());
         setPreferredSize(new Dimension(550, 650));
+
         contentPane = new JPanel();
         contentPane.setLayout(new BoxLayout(contentPane, BoxLayout.Y_AXIS));
         contentPane.add(Box.createVerticalStrut(15));
-//        contentPane.setBackground(Color.cyan);
+
         for (VideoDTO dto : dtos) {
             VideoInfoCell cell = new VideoInfoCell(dto);
             contentPane.add(cell);
-            //셀간 간격
             contentPane.add(Box.createVerticalStrut(15));
         }
 
         scrollPane = new JScrollPane(contentPane);
         scrollPane.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_ALWAYS);
-//        scrollPane.setBackground(Color.BLUE);
 
         add(scrollPane, BorderLayout.CENTER);
     }
 
     public void updatePanel(ArrayList<VideoDTO> dtos) {
-        contentPane.removeAll(); // 이전에 추가된 모든 컴포넌트 제거
+        contentPane.removeAll();
 
         for (VideoDTO dto : dtos) {
             VideoInfoCell cell = new VideoInfoCell(dto);
             contentPane.add(cell);
-            // 셀간 간격
             contentPane.add(Box.createVerticalStrut(15));
         }
 
-        contentPane.revalidate(); // 패널을 다시 그리기 위해 호출
+        contentPane.revalidate();
         contentPane.repaint();
 
-        //스크롤바 최상단 재설정
         SwingUtilities.invokeLater(() -> {
             JScrollBar verticalScrollBar = scrollPane.getVerticalScrollBar();
             verticalScrollBar.setValue(0);
