+++ client/src/controller/ViewController.java
@@ -1,63 +1,48 @@
 package controller;
 
-import GUI.Download.*;
-import GUI.Login.LoginSuperFrame;
+import GUI.Auth.LoginSuperFrame;
+import GUI.Common.CompNames;
+import GUI.Download.DownloadSuperFrame;
+import GUI.Download.DownloadWaitingPanel;
+import GUI.Download.DownloadedListPanel;
+import GUI.Download.VideoSearchPanel;
 import dto.UserDTO;
 import dto.VideoDTO;
 import service.CrawlService;
 import service.ServerService;
 
 import javax.swing.*;
-import javax.swing.text.View;
 import java.awt.*;
 import java.io.IOException;
 import java.io.ObjectOutputStream;
 import java.net.Socket;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.Objects;
-
+import java.util.*;
 
 public class ViewController {
-    // settings
-    Socket socket;
-    Thread recvThread;
-    Thread sendThread;
     public static String access = null;
     public static JFrame loginFrame = null;
     public static JFrame downFrame = null;
-//    public static JFrame sampleFrame;
-
     public static ObjectOutputStream clientOut;
     public static ServerService serverService;
-    public static CrawlService crawlService;
-
-    public static TestYoutubeService testYoutubeService = new TestYoutubeService();
-    public static TestYoutubeService2 testYoutubeService2 = new TestYoutubeService2();
-
-    //검색용 리스트
-    //장바구니 리스트
-    //다운로드 리스트
-
-    //검색 결과를 담아서 셀을 그릴 용도의 리스트
+    public static CrawlService testYoutubeService2 = new CrawlService();
+    //TODO: 검색, 다운대기, 다운로드용 리스트
     public static ArrayList<VideoDTO> videoSearchList = new ArrayList<>();
-
     public static LinkedList<VideoDTO> downloadWaitingList = new LinkedList<>();
-
     public static ArrayList<VideoDTO> downloadedList = new ArrayList<>();
+    // TODO: 소켓 통신 settings
+    Socket socket;
+    Thread recvThread;
+    Thread sendThread;
 
-    // constructor
+    //TODO: 뷰컨 생성자
     public ViewController(Socket socket) throws Exception {
-
         this.socket = socket;
         clientOut = new ObjectOutputStream(socket.getOutputStream());
         /** frame setting **/
         loginFrame = new LoginSuperFrame();
         addAuthListener();
-//        sampleFrame = new SampleFrame();
         /** get service **/
         serverService = new ServerService();
-        crawlService = new CrawlService();
         /** recv thread open **/
         this.recvThread = new Thread(new RecvController(socket));
         recvThread.start();
@@ -66,20 +51,21 @@ public class ViewController {
         sendThread.start();
     }
 
-    // functions
+    //TODO: 화면 전환
     public static void switcher() {
-        if(downFrame == null) {
+        if (downFrame == null) {
             downFrame = new DownloadSuperFrame();
             addDownloadListener();
 
-            JPanel videoSearchPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.videoSearchPanel_l);
-            ArrayList<VideoDTO> initData = testYoutubeService2.searchYoutubeVideos("춘식이");
+
+            //TODO: 초기 화면 검색어 (랜덤하게 나옴)
+            JPanel videoSearchPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.videoSearchPanel_l);
+            ArrayList<VideoDTO> initData = testYoutubeService2.searchYoutubeVideos(getRandomSong());
             ((VideoSearchPanel) videoSearchPanel).updatePanel(initData);
 
             loginFrame.dispose();
             loginFrame = null;
-        }
-        else if(loginFrame == null) {
+        } else if (loginFrame == null) {
             loginFrame = new LoginSuperFrame();
             addAuthListener();
             downFrame.dispose();
@@ -87,15 +73,15 @@ public class ViewController {
         }
     }
 
-    // Auth Frame의 컴포넌트에 액션 넣기
-    public static void addAuthListener(){
+    // TODO: 로그인 화면 액션 리스너
+    public static void addAuthListener() {
         // all components
-        JTextField idTextField = ((JTextField)ViewController.findComponentByName(loginFrame.getContentPane(), "idTextField"));
-        JTextField passwordField = ((JTextField)ViewController.findComponentByName(loginFrame.getContentPane(), "passwordField"));
-        JButton loginButton = ((JButton)ViewController.findComponentByName(loginFrame.getContentPane(), "loginButton"));
-        JButton signUpButton = ((JButton)ViewController.findComponentByName(loginFrame.getContentPane(), "signUpButton"));
-        // signUpButton
-        signUpButton.addActionListener(e->{
+        JTextField idTextField = ((JTextField) ViewController.findComponentByName(loginFrame.getContentPane(), "idTextField"));
+        JTextField passwordField = ((JTextField) ViewController.findComponentByName(loginFrame.getContentPane(), "passwordField"));
+        JButton loginButton = ((JButton) ViewController.findComponentByName(loginFrame.getContentPane(), "loginButton"));
+        JButton signUpButton = ((JButton) ViewController.findComponentByName(loginFrame.getContentPane(), "signUpButton"));
+        // TODO: signUpButton
+        signUpButton.addActionListener(e -> {
             System.out.println("Click signUpButton");
             try {
                 UserDTO dto = new UserDTO();
@@ -106,8 +92,8 @@ public class ViewController {
                 throw new RuntimeException(ex);
             }
         });
-        // loginButton
-        loginButton.addActionListener(e->{
+        // TODO: loginButton
+        loginButton.addActionListener(e -> {
             System.out.println("Click loginButton");
             try {
                 UserDTO dto = new UserDTO();
@@ -120,20 +106,21 @@ public class ViewController {
         });
     }
 
-    // Download Frame의 컴포넌트에 액션 넣기
-    public static void addDownloadListener(){
-        //TODO: 위 패널
-        JButton logOutButton = ((JButton)ViewController.findComponentByName(downFrame.getContentPane(), DownloadCompNames.logOutButton_u));
-        JTextField searchField = (JTextField)ViewController.findComponentByName(downFrame.getContentPane(), DownloadCompNames.searchField_u);
-        JButton searchButton = (JButton)ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.searchButton_u);
-        JPanel videoSearchPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.videoSearchPanel_l);
-        JButton addToDownloadButton = (JButton) ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.addToDownloadButton_l);
-        JPanel downloadWaitingPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.downloadWaitingPanel_r);
-        JButton downloadButton = (JButton) ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.downloadButton_r);
-        JPanel downloadedListPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.downloadedListPanel_r);
+    // TODO 다운로드 화면 리스너
+    public static void addDownloadListener() {
+        JButton logOutButton = ((JButton) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.logOutButton_u));
+        JTextField searchField = (JTextField) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.searchField_u);
+        JButton searchButton = (JButton) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.searchButton_u);
+        JPanel videoSearchPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.videoSearchPanel_l);
+        JButton addToDownloadButton = (JButton) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.addToDownloadButton_l);
+        JPanel downloadWaitingPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.downloadWaitingPanel_r);
+        JButton downloadButton = (JButton) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.downloadButton_r);
+        JPanel downloadedListPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.downloadedListPanel_r);
+        JPanel downloadedListCell = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.downloadedListCell_r);
+        JButton playButton = (JButton) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.downloadButton_r);
 
-        // TODO: 상 단 패 널
-        logOutButton.addActionListener(e->{
+        // TODO: 로그아웃 버튼 (여분 버튼은 숨겨뒀습니다.)
+        logOutButton.addActionListener(e -> {
             System.out.println("Click logOutButton");
             try {
                 UserDTO dto = new UserDTO();
@@ -144,16 +131,13 @@ public class ViewController {
             }
         });
 
-        //TODO: 좌측 패널
+        //TODO: 검색 텍스트 필드
         searchButton.addActionListener(e -> {
             String songName = searchField.getText();
 
             SwingWorker<ArrayList<VideoDTO>, Void> worker = new SwingWorker<>() {
                 @Override
                 protected ArrayList<VideoDTO> doInBackground() throws Exception {
-                    // 백그라운드에서 실행될 작업 수행
-//                    return testYoutubeService.searchAndDisplayResults(songName);
-//                    return testYoutubeService2.searchYoutubeVideos(songName);
                     return testYoutubeService2.searchYoutubeVideos2(songName);
                 }
 
@@ -179,12 +163,17 @@ public class ViewController {
             SwingWorker<Void, Void> worker = new SwingWorker<>() {
                 @Override
                 protected Void doInBackground() {
-                    for (VideoDTO dto : downloadWaitingList) {
+                    Iterator<VideoDTO> iterator = downloadWaitingList.iterator();
+                    while (iterator.hasNext()) {
+                        VideoDTO dto = iterator.next();
                         boolean result = testYoutubeService2.downloadWithYoutubeDL(dto.getUrl());
                         if (result) {
                             System.out.println("다운로드 성공!");
-                            downloadWaitingList.remove(dto);
+                            iterator.remove();
                             downloadedList.add(dto);
+
+                            ((DownloadWaitingPanel) downloadWaitingPanel).updatePanel(downloadWaitingList);
+                            ((DownloadedListPanel) downloadedListPanel).updatePanel(downloadedList);
                         } else {
                             System.out.println("다운로드 실패 ㅜ");
                         }
@@ -194,35 +183,19 @@ public class ViewController {
 
                 @Override
                 protected void done() {
-                    // 작업이 완료되면 UI 업데이트 및 다운로드 버튼 활성화
-                    ((DownloadWaitingPanel) downloadWaitingPanel).updatePanel(downloadWaitingList);
-                    ((DownloadedListPanel) downloadedListPanel).updatePanel(downloadedList);
                     downloadButton.setEnabled(true);
                 }
             };
 
-            // SwingWorker 실행
             worker.execute();
         });
 
-
-
-        //TODO: 우 하단 패널
-
-
     }
-
-//    public static void updateDownloadWaitingPanel() {
-////        ((VideoSearchPanel) videoSearchPanel).updatePanel(videoSearchList);
-//
-//        JPanel downloadWaitingPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(),DownloadCompNames.downloadWaitingPanel_r);
-//        ((DownloadWaitingPanel) downloadWaitingPanel).updatePanel(downloadWaitingList);
-//    }
 
     public static void updateDownloadWaitingPanel() {
         System.out.println("대기 패널 업데이트를 시작합니다.");
         SwingWorker<Void, Void> panelUpdater = new SwingWorker<Void, Void>() {
-            JPanel downloadWaitingPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(), DownloadCompNames.downloadWaitingPanel_r);
+            final JPanel downloadWaitingPanel = (JPanel) ViewController.findComponentByName(downFrame.getContentPane(), CompNames.downloadWaitingPanel_r);
 
             @Override
             protected Void doInBackground() throws Exception {
@@ -232,8 +205,6 @@ public class ViewController {
         };
         panelUpdater.execute();
     }
-
-
 
     // 이름으로 컴포넌트 찾기 메서드
     public static Component findComponentByName(Container container, String componentName) {
@@ -252,28 +223,9 @@ public class ViewController {
         return null; // 이름이 일치하는 컴포넌트가 없는 경우
     }
 
-    // 이름으로 컴포넌트 찾기 메서드
-//    public static Component findComponentBySameName(Container container, String componentName) {
-//        System.out.println("______________________");
-//        Component[] components = container.getComponents();
-//        System.out.println("test222222");
-//        for (Component component : components) {
-//            System.out.println("for" + component.toString());
-//            if (Objects.equals(component.getName(), componentName)) {
-//                System.out.println("______________________");
-//                ((AddToDownloadButton)component).addActionListener(e->{
-//                    downloadWaitingList.add(((AddToDownloadButton)component).getDto());
-//                    System.out.println("끄아악");
-//                });
-//            }
-////            if (component instanceof Container) {
-////                Component foundComponent = findComponentByName((Container) component, componentName);
-////                if (foundComponent != null) {
-////                    return foundComponent; // 하위 컨테이너에서 찾은 컴포넌트 반환
-////                }
-////            }
-//        }
-//        System.out.println("없음");
-//        return null; // 이름이 일치하는 컴포넌트가 없는 경우
-//    }
+    public static String getRandomSong() {
+        String[] initSongs = {"데이식스", "아이유", "쏜애플", "백예린"};
+        Random random = new Random();
+        return initSongs[random.nextInt(initSongs.length)];
+    }
 }
