+++ client/src/service/CrawlService.java
@@ -1,10 +1,6 @@
 package service;
 
 import dto.VideoDTO;
-import controller.ViewController;
-import static controller.ViewController.downFrame;
-
-
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
@@ -14,123 +10,42 @@ import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
-import javax.imageio.ImageIO;
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.image.BufferedImage;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.UnsupportedEncodingException;
-import java.net.URL;
-import java.net.URLEncoder;
+import java.io.*;
 import java.text.DecimalFormat;
-
+import java.util.ArrayList;
 
 public class CrawlService {
-    // settings
+    private static volatile int downloadProgress = 0; // volatile 키워드 추가
 
-    // constructor
-    public CrawlService(){
-        VideoDTO videoDto = new VideoDTO();
+    public static synchronized void incrementDownloadProgress() {
+        downloadProgress++;
     }
 
-    // functions
-    public void searchAndDisplayResults() {
-        JTextField songNameField = ((JTextField)ViewController.findComponentByName(downFrame.getContentPane(), ""));
-        JTextArea resultArea = (JTextArea)ViewController.findComponentByName(downFrame.getContentPane(), "");
-        JPanel videoPanel = (JPanel)ViewController.findComponentByName(downFrame.getContentPane(), "");
+    public static int getDownloadProgress() {
+        return downloadProgress;
+    }
 
-        if (songNameField == null || resultArea == null || videoPanel == null){
-            return;
-        }
-        String songName = songNameField.getText();
-        resultArea.setText(""); // 이전 결과 지우기
-        videoPanel.removeAll(); // 이전 영상 정보 패널 지우기
+    public static void resetDownloadProgress() {
+        downloadProgress = 0;
+    }
 
+    public ArrayList<VideoDTO> searchYoutubeVideos(String name) {
+
+        ArrayList<VideoDTO> videoList = new ArrayList<>();
+        String songName = name;
         try {
-            String encodedSongName = URLEncoder.encode(songName, "UTF-8");
-            String[] videoUrls = searchYoutubeVideos(encodedSongName);
-
-            if (videoUrls != null) {
-                for (String videoUrl : videoUrls) {
-                    // 동영상 정보 가져오기
-                    String videoInfoUrl = "https://www.youtube.com/oembed?url=" + videoUrl + "&format=json";
-                    Document videoInfoDoc = Jsoup.connect(videoInfoUrl).ignoreContentType(true).get();
-                    String videoInfoJson = videoInfoDoc.body().text();
-                    JSONParser parser = new JSONParser();
-                    JSONObject videoInfoObject = (JSONObject) parser.parse(videoInfoJson);
-
-                    // 영상 제목
-                    String title = (String) videoInfoObject.get("title");
-                    // 조회수
-                    Long viewCount = getViewCount(videoUrl);
-                    System.out.println(viewCount);
-                    String formattedViewCount = formatViewCount(viewCount);
-                    // 업로드한 사람
-                    String uploader = (String) videoInfoObject.get("author_name");
-                    // videoUrl에서 VIDEO_ID 추출
-                    String videoId = videoUrl.substring(videoUrl.indexOf("v=") + 2);
-                    // 썸네일 이미지 URL 조합
-                    String thumbnailUrl = "https://img.youtube.com/vi/" + videoId + "/0.jpg";
-
-                    // 영상 정보와 썸네일 이미지를 VideoInfo 객체로 묶음
-                    VideoDTO videoInfo = new VideoDTO();
-                    videoInfo.setTitle(title);
-                    videoInfo.setViewCount(formattedViewCount);
-                    videoInfo.setThumbnailUrl(thumbnailUrl);
-                    videoInfo.setUploader(uploader);
-                    videoInfo.setUrl(videoUrl);
-                    // VideoInfo 객체를 사용하여 영상 정보 패널 생성
-                    createVideoPanel(videoInfo);
-                }
-            } else {
-                resultArea.append("해당 곡을 찾을 수 없습니다.");
-            }
-
-            // 패널 갱신
-            videoPanel.revalidate();
-            videoPanel.repaint();
-
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
-            JOptionPane.showMessageDialog(downFrame, "곡 이름을 인코딩하는 중 오류가 발생했습니다.");
-        } catch (IOException | ParseException e) {
-            e.printStackTrace();
-        }
-    }
-    public static long getViewCount(String youtubeUrl) throws IOException {
-        Document doc = Jsoup.connect(youtubeUrl).get();
-        // 유튜브의 조회수는 <meta> 태그에 og:description 속성으로 들어가 있음
-        Element metaTag = doc.select("meta[itemprop=interactionCount]").first();
-        if (metaTag != null) {
-            String viewCountText = metaTag.attr("content");
-            return Long.parseLong(viewCountText);
-        } else {
-            throw new IOException("조회수를 찾을 수 없습니다.");
-        }
-    }
-
-    public static String formatViewCount(long viewCount) {
-        DecimalFormat formatter = new DecimalFormat("#,###");
-        return formatter.format(viewCount);
-    }
-
-    public String[] searchYoutubeVideos(String encodedSongName) {
-        try {
-            String url = "https://www.youtube.com/results?search_query=" + encodedSongName;
+            String url = "https://www.youtube.com/results?search_query=" + songName;
             Document doc = Jsoup.connect(url).get();
-            Elements elements = doc.select("ytInitialData");
+            Elements elements = doc.select("script");
 
-            String[] videoUrls = new String[10]; // 최대 10개의 비디오 URL 저장
             int count = 0; // URL 개수 카운트
 
-            for (Element script : doc.getElementsByTag("script")) {
+            for (Element script : elements) {
                 String scriptContent = script.html();
                 if (scriptContent.contains("ytInitialData")) {
-                    String jsonData = scriptContent.substring(scriptContent.indexOf("{"), scriptContent.lastIndexOf("}") + 1);
+                    int startIndex = scriptContent.indexOf("ytInitialData");
+                    int endIndex = scriptContent.lastIndexOf("};") + 1;
+                    String jsonData = scriptContent.substring(startIndex + 16, endIndex);
                     JSONParser parser = new JSONParser();
                     JSONObject jsonObject = (JSONObject) parser.parse(jsonData);
 
@@ -139,6 +54,7 @@ public class CrawlService {
                             .get("primaryContents"))
                             .get("sectionListRenderer"))
                             .get("contents");
+
                     JSONObject contentsObject = (JSONObject) contentsArray.get(0);
                     JSONArray result = (JSONArray) ((JSONObject) contentsObject.get("itemSectionRenderer")).get("contents");
 
@@ -148,110 +64,222 @@ public class CrawlService {
 
                         if (videoRenderer != null && videoRenderer.containsKey("videoId")) {
                             String videoId = (String) videoRenderer.get("videoId");
+                            String title = ((JSONObject)  ((JSONArray) ((JSONObject)videoRenderer.get("title")).get("runs")).get(0)).get("text").toString();
+                            String uploader = ((JSONObject) ((JSONArray) ((JSONObject) videoRenderer.get("ownerText")).get("runs")).get(0)).get("text").toString();
+
+                            Long viewCount = Long.parseLong(((JSONObject) videoRenderer.get("viewCountText")).get("simpleText").toString().replaceAll("[^0-9]", ""));
+                            String thumbnailUrl = "https://img.youtube.com/vi/" + videoId + "/0.jpg";;
                             String videoUrl = "https://www.youtube.com/watch?v=" + videoId;
-                            videoUrls[count++] = videoUrl;
+                            VideoDTO videoInfo = new VideoDTO();
+                            videoInfo.setTitle(title);
+                            videoInfo.setViewCount(formatViewCount(viewCount));
+                            videoInfo.setUploader(uploader);
+                            videoInfo.setThumbnailUrl(thumbnailUrl);
+                            videoInfo.setUrl(videoUrl);
+
+                            videoList.add(videoInfo);
+                            count++;
 
                             if (count >= 10) {
-                                return videoUrls;
+                                break;
                             }
                         }
                     }
                     break;
                 }
             }
+        } catch (IOException | ParseException e) {
+            e.printStackTrace();
+        }
+        return videoList;
+    }
 
-            if (!elements.isEmpty()) {
-                for (Element firstVideo : elements) {
-                    String videoUrl = "https://www.youtube.com" + firstVideo.attr("href");
-                    videoUrls[count++] = videoUrl;
+    public ArrayList<VideoDTO> searchYoutubeVideos2(String encodedSongName) {
+        ArrayList<VideoDTO> videoList = new ArrayList<>();
 
-                    if (count >= 10) {
-                        return videoUrls;
+        try {
+            String url = "https://www.youtube.com/results?search_query=" + encodedSongName;
+            Document doc = Jsoup.connect(url).get();
+            Elements elements = doc.select("script");
+
+            int count = 0; // URL 개수 카운트
+
+            for (Element script : elements) {
+                String scriptContent = script.html();
+                if (scriptContent.contains("ytInitialData")) {
+                    int startIndex = scriptContent.indexOf("ytInitialData");
+                    int endIndex = scriptContent.lastIndexOf("};") + 1;
+                    String jsonData = scriptContent.substring(startIndex + 16, endIndex);
+                    JSONParser parser = new JSONParser();
+                    JSONObject jsonObject = (JSONObject) parser.parse(jsonData);
+
+                    JSONArray contentsArray = (JSONArray) ((JSONObject) ((JSONObject) ((JSONObject) ((JSONObject) jsonObject.get("contents"))
+                            .get("twoColumnSearchResultsRenderer"))
+                            .get("primaryContents"))
+                            .get("sectionListRenderer"))
+                            .get("contents");
+
+                    for (Object content : contentsArray) {
+                        JSONObject contentObject = (JSONObject) content;
+                        if (contentObject.containsKey("itemSectionRenderer")) {
+                            JSONArray itemArray = (JSONArray) ((JSONObject)contentObject.get("itemSectionRenderer")).get("contents");
+                            for (Object obj : itemArray) {
+                                JSONObject item = (JSONObject) obj;
+                                JSONObject videoRenderer = (JSONObject) item.get("videoRenderer");
+
+                                if (videoRenderer != null && videoRenderer.containsKey("videoId")) {
+                                    String videoId = (String) videoRenderer.get("videoId");
+                                    String title = ((JSONObject) ((JSONArray) ((JSONObject) videoRenderer.get("title")).get("runs")).get(0)).get("text").toString();
+                                    String uploader = ((JSONObject) ((JSONArray) ((JSONObject) videoRenderer.get("ownerText")).get("runs")).get(0)).get("text").toString();
+
+                                    Long viewCount = Long.parseLong(((JSONObject) videoRenderer.get("viewCountText")).get("simpleText").toString().replaceAll("[^0-9]", ""));
+                                    String thumbnailUrl = "https://img.youtube.com/vi/" + videoId + "/0.jpg";
+                                    String videoUrl = "https://www.youtube.com/watch?v=" + videoId;
+                                    VideoDTO videoInfo = new VideoDTO();
+                                    videoInfo.setTitle(title);
+                                    videoInfo.setViewCount(formatViewCount(viewCount));
+                                    videoInfo.setUploader(uploader);
+                                    videoInfo.setThumbnailUrl(thumbnailUrl);
+                                    videoInfo.setUrl(videoUrl);
+
+                                    videoList.add(videoInfo);
+                                    count++;
+
+                                    if (count >= 10) {
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+                        if (count >= 10) {
+                            break;
+                        }
                     }
+                    break;
                 }
             }
         } catch (IOException | ParseException e) {
             e.printStackTrace();
         }
-        return null;
+        return videoList;
     }
 
 
-    private void createVideoPanel(VideoDTO videoDto) {
-        JPanel videoInfoPanel = new JPanel(new BorderLayout());
 
-        // 썸네일 이미지를 표시하기 위해 JLabel을 생성하여 패널에 추가
-        JLabel thumbnailLabel = new JLabel(new ImageIcon(getResizedImage(videoDto.getThumbnailUrl(), 120, 90)));
-        videoInfoPanel.add(thumbnailLabel, BorderLayout.WEST);
 
-        // 영상 정보를 텍스트 영역에 표시하기 위해 JTextArea를 생성하여 패널에 추가
-        JTextArea textArea = new JTextArea();
-        textArea.append("Title: " + videoDto.getTitle() + "\n");
-        textArea.append("View Count: " + videoDto.getViewCount() + "\n");
-        textArea.append("Uploader: " + videoDto.getUploader() + "\n");
-        textArea.append("Video URL: " + videoDto.getUrl() + "\n");
-        textArea.setEditable(false); // 수정 불가능하도록 설정
-        videoInfoPanel.add(textArea, BorderLayout.CENTER);
 
-        // 다운로드 버튼 생성 및 패널에 추가
-        JButton downloadButton = new JButton("다운로드");
-        downloadButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                downloadWithYoutubeDL(videoDto.getUrl()); // 다운로드 메서드 호출
-            }
-        });
-        videoInfoPanel.add(downloadButton, BorderLayout.EAST); // EAST로 배치
 
-        // 영상 정보 패널을 전체 패널에 추가
-        JPanel videoPanel = (JPanel)ViewController.findComponentByName(downFrame.getContentPane(), "");
-        if (videoPanel == null){
-            return;
-        }
-        videoPanel.add(videoInfoPanel);
+
+    private String formatViewCount(Long viewCount) {
+        DecimalFormat decimalFormat = new DecimalFormat("#,###");
+        String formattedCount = decimalFormat.format(viewCount);
+        return formattedCount + "회";
     }
 
-    // 이미지를 리사이즈하는 메서드
-    public Image getResizedImage(String imageUrl, int width, int height) {
-        try {
-            URL url = new URL(imageUrl);
-            BufferedImage originalImage = ImageIO.read(url);
-            Image resizedImage = originalImage.getScaledInstance(width, height, Image.SCALE_SMOOTH);
-            return resizedImage;
-        } catch (IOException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    public void downloadWithYoutubeDL(String videoUrl) {
+    public boolean downloadWithYoutubeDL(String videoUrl) {
         try {
             // youtube-dlp의 전체 경로 지정
-            String youtubeDLPPath = "C:\\Users\\TECH2-16\\Desktop\\java pjt\\client\\yt-dlp";
+            String rootDirectory = System.getProperty("user.dir");
+            String youtubeDLPPath = rootDirectory + "/yt-dlp";
+
+            //TODO: 저장 경로
+//            File downloadList = new File("C:\\Users\\김병준\\Desktop\\video-downloader\\client\\downloadList");
+            File downloadList = new File(rootDirectory+"/downloadList");
+//            File downloadList = new File(System.getProperty("user.dir") + "\\client\\downloadList");
+
+            if (!downloadList.exists()) {
+                downloadList.mkdirs(); // 디렉토리가 존재하지 않으면 생성합니다.
+            }
             // youtube-dlp 실행 명령어와 옵션을 따로 분리하여 전달
             // 직접 URL을 사용하여 다운로드
             String[] command = {"cmd", "/c", youtubeDLPPath + ".exe", "-f", "bestvideo+bestaudio/best", "--merge-output-format", "mp4", "-o", "%(title)s.mp4", videoUrl};
 
             ProcessBuilder builder = new ProcessBuilder(command);
-            builder.redirectErrorStream(true);
-            Process process = builder.start();
 
+            builder.redirectErrorStream(true);
+
+            builder.directory(downloadList);
+
+            Process process = builder.start();
             // 이하 코드는 그대로 유지
             BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
             String line;
             while ((line = reader.readLine()) != null) {
+                System.out.println("이게 다운 속도인가?");
+                downloadProgress++;
                 System.out.println(line);
+                System.out.println("서비스의 진행도: "+ downloadProgress);
             }
 
             int exitCode = process.waitFor();
             if (exitCode == 0) {
-                JOptionPane.showMessageDialog(downFrame, "다운로드가 완료되었습니다.");
+                downloadProgress = 0;
+                return true;
             } else {
-                JOptionPane.showMessageDialog(downFrame, "다운로드 중 오류가 발생했습니다.");
+                downloadProgress = 0;
+                return false;
             }
         } catch (IOException | InterruptedException ex) {
             ex.printStackTrace();
-            JOptionPane.showMessageDialog(downFrame, "다운로드 중 오류가 발생했습니다.");
+            downloadProgress = 0;
+            return false;
         }
     }
+
+
+    public static void playDownloadedVideo(VideoDTO dto) {
+        // 다운로드된 영상 파일의 경로 설정
+        String name = dto.getTitle();
+        System.out.println("타이틀: "+name);
+
+        String rootDirectory = System.getProperty("user.dir");
+
+        String filePath = rootDirectory+"\\downloadList\\"+name+".mp4";
+        System.out.println("파일 경로: "+filePath);
+//        File file = new File(filePath);
+//        Desktop desktop = Desktop.getDesktop();
+//        try {
+//            desktop.open(file);
+//
+//        }catch (Exception e) {
+//            System.out.println(e.getLocalizedMessage());
+//            e.printStackTrace();
+//        }
+
+        System.out.println("파일 경로입니다아아아아"+filePath);
+
+        try {
+//            ProcessBuilder builder = new ProcessBuilder("cmd.exe", "/c", "start", filePath);
+            ProcessBuilder builder = new ProcessBuilder(filePath);
+            builder.start();
+        } catch (IOException ex) {
+            ex.printStackTrace();
+
+        }
+    }
+
+//    downloadButton.addActionListener(new ActionListener() {
+//        @Override
+//        public void actionPerformed(ActionEvent e) {
+//            downloadVideo();
+//        }
+//    });
+
 }
+
+//C:\Users\김병준\Desktop\video-downloader\client\downloadListKung Ddari Sha Bah Rah (꿍따리 샤바라).mp4
+//C:\Users\김병준\Desktop\video-downloader\client\downloadListKung Ddari Sha Bah Rah (꿍따리 샤바라).mp4
+
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList\"춘식이 이모티콘 무료라서 받았는데" 항의 폭주...카카오는 해명 [지금이뉴스] / YTN.mp4
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList\"춘식이 이모티콘 무료라서 받았는데" 항의 폭주...카카오는 해명 [지금이뉴스] \ YTN.mp4 doesn't exist.
+
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList/"춘식이 이모티콘 무료라 받았는데" 항의 폭주…카카오 "실수" / SBS / 실시간 e뉴스.mp4
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList\"춘식이 이모티콘 무료라 받았는데" 항의 폭주…카카오 "실수" \ SBS \ 실시간 e뉴스.mp4 doesn't exist.
+
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList\"춘식이 이모티콘 무료라서 받았는데" 항의 폭주...카카오는 해명 [지금이뉴스] / YTN.mp4
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList\"춘식이 이모티콘 무료라서 받았는데" 항의 폭주...카카오는 해명 [지금이뉴스] \ YTN.mp4
+
+// filePath 경로
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList\Kung Ddari Sha Bah Rah (꿍따리 샤바라).mp4
+//
+//C:\Users\김병준\Desktop\video-downloader\client\downloadList\Kung Ddari Sha Bah Rah (꿍따리 샤바라).mp4
